% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{assertColors}
\alias{assertColors}
\title{INTERNAL FUNCTION: Fixes colors to include random}
\usage{
assertColors(colorsIn)
}
\arguments{
\item{colorsIn}{List of colors}
}
\value{
List of colors, size might be greater, if random has a number > 1
}
\description{
Goes through the list of colors, and handles any random entries, specifically if a number is provided
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
