% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{confirmStatus}
\alias{confirmStatus}
\title{Loads content is check if page is valid}
\usage{
confirmStatus(
  source,
  url = paste0(paste0("https://www.uniprot.org/uniprot/", source), ".xml"),
  toRunIf,
  toRunElse
)
}
\arguments{
\item{source}{Protein UniProt keycode}

\item{url}{URL source to search}

\item{toRunIf}{Function with passed in webpage content, should the page be valid}

\item{toRunElse}{Function to run should the page be invalid}
}
\value{
Function specific output
}
\description{
Will attempt to see if the source is a valid uniprot page. Will return a run functions
depending if the code is valid (code == 200) or invalid
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
