% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{dp}
\alias{dp}
\title{Interanl function that can generate a plot for one protein}
\usage{
dp(data, height = 1.3)
}
\arguments{
\item{data}{Protein's dataFrame}

\item{height}{The height of the plot}
}
\value{
If dataFrame provided is of the correct format, will return a blank plot
}
\description{
This function will take a protein's dataFrame and make a plot from it
}
\examples{
xmls <- getProteinRemote(c("Q04206.xml", "Q9D270.xml"))
features <- getFeatureList(xmls)
data <- featuresToDataFrame(features)
plotForSingleProtein <- dp(data[[1]])

}
