% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{drawChain}
\alias{drawChain}
\title{INTERNAL FUNCTION: Main draw function}
\usage{
drawChain(figure, xi, xf, yi, yf, info, clr, offset = 0)
}
\arguments{
\item{figure}{Main plotly plot/figure}

\item{xi}{Starting location for drawing on X-axis}

\item{xf}{Ending location for drawing on the X-axis}

\item{yi}{Starting location for drawing on the Y-axis}

\item{yf}{Ending location for drawing on the Y-axis}

\item{info}{Description info, used for legend nameing and hover info}

\item{clr}{Color}

\item{offset}{Offset on the X-axis of where to start drawing}
}
\value{
Returns the plot, with the new feature drawn on
}
\description{
Used to add the the feature to the figure
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
