% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{drawDomains}
\alias{drawDomains}
\title{Add all domains from protein's features to respective plot}
\usage{
drawDomains(plots, data)
}
\arguments{
\item{plots}{List of protein plots}

\item{data}{List of protein feature dataFrames}
}
\value{
A list of plots for each of same size as provided, with modified values for each element
}
\description{
This function is a utility function, that will interate through lists of proteins dataFrames and find all domains.
It will then add the domain to the respetive plot. Will return original list of plots with modified values for the
contained plots
}
\examples{
xmls <- getProteinRemote(c("Q04206.xml", "Q9D270.xml"))
features <- getFeatureList(xmls)
data <- featuresToDataFrame(features)
plots <- drawPlot(data)
plots <- drawDomains(plots, data)

}
