% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{drawFeature}
\alias{drawFeature}
\title{INTERNAL FUNCTION: Helper function for standardizing input for before drawing}
\usage{
drawFeature(figure, d, toParse, condition, yStart, yStop, offset = 0)
}
\arguments{
\item{figure}{Main plotly plot/figure}

\item{d}{Dataframe containing proteins' features}

\item{toParse}{Specific entires that will search through the features
dataframe to find matchs based upon condition}

\item{condition}{Condition function for when to return true for match}

\item{yStart}{Height element should begin drawing from}

\item{yStop}{Height element should stop drawing at}

\item{offset}{Offset of where drawing on the X-axis takes place}
}
\value{
Returns figure will all matching features drawn on
}
\description{
Corrects input varables to proper form and used to interate through entries
that were found based upon input condition.
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
