% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{drawPlot}
\alias{drawPlot}
\title{Generate a plot for a list of protein dataFrames}
\usage{
drawPlot(dataList)
}
\arguments{
\item{dataList}{List of protein dataFrames}
}
\value{
A list of plots for each protein feature dataFrame
}
\description{
This function will take a list of protein dataFrames and generate a list of plots for each.
}
\examples{
xmls <- getProteinRemote(c("Q04206.xml", "Q9D270.xml"))
features <- getFeatureList(xmls)
data <- featuresToDataFrame(features)
plots <- drawPlot(data)

}
