% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{elementIfMatch}
\alias{elementIfMatch}
\title{Find all elements from features list that matches pattern}
\usage{
elementIfMatch(plots, data, type, xoff = 0, ymi = 0, yma = 1)
}
\arguments{
\item{plots}{List of protein plots}

\item{data}{List of protein feature dataFrames}

\item{type}{Pattern used for matching}

\item{xoff}{Offset for the bar, defines how much thickness is added}

\item{ymi}{The Y value where drawing will start}

\item{yma}{The Y value where drawing will stop}
}
\value{
A list of plots for each of same size as provided, with modified values for each element
}
\description{
This function will interate through lists of proteins dataFrames and find all elements that match
type pattern. It will then add the element to the respetive plot. Will return original list of plots
with modified values for the contained plots
}
\examples{
xmls <- getProteinRemote(c("Q04206.xml", "Q9D270.xml"))
features <- getFeatureList(xmls)
data <- featuresToDataFrame(features)
plots <- drawPlot(data)
plots <- elementIfMatch(plots, data, "chain")

}
