% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataParsing.R
\name{featuresToDataFrame}
\alias{featuresToDataFrame}
\title{Convert a list of feature list for each protein into a dataFrame}
\usage{
featuresToDataFrame(featuresList)
}
\arguments{
\item{featuresList}{A list of feature lists for each loaded protein}
}
\value{
A list of dataFrames for each individual protein
}
\description{
This function will take a list, where each entry represent the data for a separate protein. It will
then interate through each sub list and look through each and add each featuers to a dataframe. Each
dataFrame generated for each proteins features are then concated together as a list of dataFrames.
}
\examples{
xmls <- getProteinRemote(c("Q04206.xml", "Q9D270.xml"))
features <- getFeatureList(xmls)
data <- featuresToDataFrame(features)

}
