% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataParsing.R
\name{getFeatureList}
\alias{getFeatureList}
\title{Wrapper method to interate a list of protein components and find all features}
\usage{
getFeatureList(proteinList)
}
\arguments{
\item{proteinList}{A list for of proteins, where each sublist contanins all of the protein's parsed elements}
}
\value{
If all provided paramters are correct, then will return a list containing sublists that are protein specific,
each containing all feature elements. The parent list size is the same size as the inputed list.
}
\description{
This function will interate through a list of proteins, and find within each protein, all feature elements
and return a list of equal size as the proteinList parameter, each sub list containing each individual protein's
features
}
\examples{
xmls <- getProteinRemote(c("Q04206.xml", "Q9D270.xml"))
features <- getFeatureList(xmls)

}
