% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataParse.R
\name{getFeaturesDataFrame}
\alias{getFeaturesDataFrame}
\title{INTERNAL FUNCTION: Creates Dataframe for features}
\usage{
getFeaturesDataFrame(xmls)
}
\arguments{
\item{xmls}{List of parsed XMLs for each protein}
}
\value{
List of dataframes for each protein's XML data
}
\description{
Cycles through all proteins and for each, finds all entries called
"features", then addes them to a dataframe for each protein and combines
all dataframes into a list
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
