% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{getLocal}
\alias{getLocal}
\title{INTERNAL FUNCTION: Loads the source from local .xml file}
\usage{
getLocal(source)
}
\arguments{
\item{source}{.xml file to load}
}
\value{
If successful, will return a list of XML data
}
\description{
INTERNAL FUNCTION: Loads the source from local .xml file
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
