% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{getProtein}
\alias{getProtein}
\title{Get protein XML data}
\usage{
getProtein(source, showProgress = TRUE)
}
\arguments{
\item{source}{List of entries for which to attempt to load}

\item{showProgress}{If the user should be notified on the progress of this function}
}
\value{
List for each loaded protein, each element containing a list of XML data
}
\description{
Searches through list or vector of entries and attempts to retrieve UniProt protein
XML data. This can be done by either .xml entry for local files, where if none are found,
will ask if the user either want to skip or download the file and load that. Otherwise, the
user can specify a UniProt protein key_code where the function will then attempt to read
the XML data from the webpage. Alternativly, the user can use "random" to randomly pick
a protein.
}
\examples{
#Example 1:
#This will get the xml data for the protein Q04206
proteinXml <- getProtein("Q04206")

#Example 2:
#This will get a random human protein and the Q04206 protein
proteinXml <- getProtein(c("random", "Q04206"))

#Example 3:
#This will get 4 random human proteins
proteinXml <- getProtein("random number:4")

#Example 4:
#This will get 3 random mouse proteins, the protein Q04206, and then 2 more random human proteins
proteinXml <- getProtein(c("random number:3 orgid:10090", "Q04206", "random number:2"))

}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
