% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{getProteinLocal}
\alias{getProteinLocal}
\title{get list of xmls loaded from local .xml files}
\usage{
getProteinLocal(fileNames)
}
\arguments{
\item{fileNames}{A file name or list of file names that have the xml format used by Uniprot}
}
\value{
If the files exist and the xml is in the same standard format Uniprot uses, then this function
will return a list of lists where the parent list is of the same size as the number of file names
provided
}
\description{
This function will interate through all file names provided and attempt to parse through it
and then convert the entry into a list, stored inside of a larger list where each
entry is a separate protein
}
\examples{
proteins <- getProteinLocal(c("Q0D270.xml","Q0D271.xml","Q0D272.xml"))

}
