% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{getProteinRemote}
\alias{getProteinRemote}
\title{Get list of xmls loaded from remote Uniprot server}
\usage{
getProteinRemote(codes)
}
\arguments{
\item{codes}{A code or list of codes of protein keys}
}
\value{
If the webpage can be found for the valid protein key, and the data from the server is able to be
retrieved, then the data will be parsed and stored as a list within a parent list of the same size as
the number of codes provided
}
\description{
This function will interate through a single or vector of provided protein keys,
load the data, parse the xml, and then store the parsed data as a list within a
parent list
}
\examples{
proteins <- getProteinRemote(c(c("Q0D270","Q0D271","Q0D272"))

}
