% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{getRandomProtein}
\alias{getRandomProtein}
\title{INTERNAL FUNCTION: Randomly picks a protein to load}
\usage{
getRandomProtein(orgID)
}
\arguments{
\item{orgID}{Organism id that UniProt uses for organism registration}
}
\value{
List of parsed XML data
}
\description{
Will use UniProts internal API function to find a random protein using the specified
organism ID
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
