% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{getRemote}
\alias{getRemote}
\title{INTERNAL FUNCTION: Load source remotly}
\usage{
getRemote(
  source,
  url = paste0(paste0("https://www.uniprot.org/uniprot/", source), ".xml")
)
}
\arguments{
\item{source}{Protein UniProt keycode}

\item{url}{URL source to search}
}
\value{
List of XML data
}
\description{
Will attempt to see if the source is a valid uniprot page. Will return a status_code to
determine if successful, and if so, parse XML data
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
