% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{getRemoteDownload}
\alias{getRemoteDownload}
\title{INTERNAL FUNCTION: Download the file and then load locally}
\usage{
getRemoteDownload(source)
}
\arguments{
\item{source}{UniProt key_code.xml to try and download .xml file}
}
\value{
Returns list of XML data
}
\description{
INTERNAL FUNCTION: Download the file and then load locally
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
