% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{getRemotePreApproved}
\alias{getRemotePreApproved}
\title{INTERNAL FUNCTION: Load data from prefound web source}
\usage{
getRemotePreApproved(get)
}
\arguments{
\item{get}{Web source content obtained using the httr library}
}
\value{
List of XML data
}
\description{
INTERNAL FUNCTION: Load data from prefound web source
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
