% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{ifelse}
\alias{ifelse}
\title{INTERNAL FUNCTION: If else function}
\usage{
ifelse(condition, true, false)
}
\arguments{
\item{condition}{Condition for which to evaluate by}

\item{true}{What value should be returned/run should the condition be true}

\item{false}{What value should be returned/run should the condition be false}
}
\value{
Value depending on should the condition be true or false
}
\description{
Made because function provided by base package A) is really slow as there
are a thousand unneeded things they do, and B) cause the function from the
base package also for some reason returns incorrect values, no clue
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
