% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataParsing.R
\name{listFromName}
\alias{listFromName}
\title{List of elements from name form a protein's component list}
\usage{
listFromName(proteinList, name)
}
\arguments{
\item{proteinList}{A list for a specific protein contaning all of its parsed elements}

\item{name}{An identifier to preform matching on}
}
\value{
If all provided paramters are correct, then will return a list of all elements with matching name
to the provided pattern name
}
\description{
This function will interate through a protein's list of components, and if any element name matches
a provided name, then will give back a list of elements. Note should be tacken, that this only will work for a
proteins specific list, not the parent list.
}
\examples{
xmls <- getProteinRemote(c("Q04206.xml", "Q9D270.xml"))
featuresForTheFirstProtein <- listFromName(xmls[[1]], "features")

}
