% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{plotProteins}
\alias{plotProteins}
\title{Draw all plots from a plot list to a single plot}
\usage{
plotProteins(plots)
}
\arguments{
\item{plots}{List of protein plots}
}
\value{
A single plot for which all plots are combine together
}
\description{
This function will take a list of plots and combine them all onto a single plot for comparison
}
\examples{
xmls <- getProteinRemote(c("Q04206.xml", "Q9D270.xml"))
features <- getFeatureList(xmls)
data <- featuresToDataFrame(features)
plots <- drawPlot(data)
plots <- drawMotifs(plots, data)
plotProteins(plots)

}
