% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{randomColor}
\alias{randomColor}
\title{Get and random rgb color}
\usage{
randomColor()
}
\value{
Returns random rgb color
}
\description{
Get and random rgb color
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
