% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUniprotProteinView.R
\name{retrieveParseData}
\alias{retrieveParseData}
\title{Retrieved data using webpage data}
\usage{
retrieveParseData(apProData)
}
\arguments{
\item{apProData}{Predownload webpage data using confirmStatus function}
}
\value{
List of XML and features dataframe
}
\description{
Will input a wepages sources, and then get all data and return the XML data and features data frame
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
