% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUniprotProteinView.R
\name{runUniprotProteinDraw}
\alias{runUniprotProteinDraw}
\title{Creates Shiny app}
\usage{
runUniprotProteinDraw()
}
\description{
Will create an interactive shiny app with imbeded plotly figure
}
\references{
Park T. Bootswatch. (2020). GitHub Repository. \href{https://github.com/thomaspark/bootswatch}{Link}

Chang, W., Cheng J., Allaire J., Xie Y., McPherson J. (2017). Shiny: web application framework for R.
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
