% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrawing.R
\name{drawProtein}
\alias{drawProtein}
\title{Drawing plotly plot using various user inputs to define what is drawn on the screen}
\usage{
drawProtein(
  proteins,
  types = list(),
  dess = list(),
  structure = list(),
  singleOffset = 1,
  title = NULL,
  saveGlobal = FALSE,
  btwnSpacingStart = 1,
  btwnSpacing = 0.3,
  showProgress = TRUE
)
}
\arguments{
\item{proteins}{A string, vector, list, or list with entry named "source" that is equal
to a string, vector or list containing valid UniProt protein key_codes. This list can also
contain an element called colors that can equal a rgb() value, string of a color, or the word
in order "random" for a random color, to specify what color the protein in the same order will
be}

\item{types}{A string, vector, list, or list with entry named "type" that is equal
to a string, vector or list containing valid matches to feature type names. This list can also
contain an element called colors that can equal a rgb() value, string of a color, or the word
in order "random" for a random color, to specify what color the element in the same order will
be}

\item{dess}{A string, vector, list, or list with entry named "type" that is equal
to a string, vector or list containing word searches that will then be used to find any entries
description, and if it should contain the string, will be valid. This list can also
contain an element called colors that can equal a rgb() value, string of a color, or the word
in order "random" for a random color, to specify what color the element in the same order will
be}

\item{structure}{A string, vector, list, or list with entry named "type" that is equal
to a string, vector or list containing valid matches to feature description names. These entries
will normally be rendered above the protein, though this can be chaged by specifiing the parameters
btwnSpacingStart and btwnSpacing. This list can also contain an element called colors that can
equal a rgb() value, string of a color, or the word in order "random" for a random color, to
specify what color the element in the same order will be}

\item{singleOffset}{The offset +- by parameter value used for rendering features specified
using the parameter dess, as these entries are noramally reserved for protein modifications,
which are single amino acids, and thus would not render otherwise, thus this parameter allows
the user to specify how thick they will show up.}

\item{title}{The title that the plot will inherit}

\item{saveGlobal}{If the function will save the protein XML data and feature dataframe
data as global values, so that the user can use them as they wish}

\item{btwnSpacingStart}{Where the elements specified in the structure parameter will start,
by default they will be rendered above the protein}

\item{btwnSpacing}{How tall the elements specified in the structure parameter will be}

\item{showProgress}{If the consol should tell the user the current status in loading the
proteins is}
}
\value{
Plotly figure
}
\description{
Creates plot using input list of protein keys for uniprot entries. Will draw main chain
then draw further features for the protein as specified within, should they be present.
Can search for features by either exact match, or if description contains an input string.
Can also further draw certain features above the structure, traditionally structural features
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
