% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataParse.R
\name{getFeaturesDataFrame}
\alias{getFeaturesDataFrame}
\title{Creates Dataframe for features}
\usage{
getFeaturesDataFrame(xmls)
}
\arguments{
\item{xmls}{List of parsed XMLs for each protein}
}
\value{
List of dataframes for each protein's XML data
}
\description{
Cycles through all proteins and for each, finds all entries called
"features", then addes them to a dataframe for each protein and combines
all dataframes into a list
}
\examples{
#Example 1:
#This will fetch the protein data, then generate the features dataframe for it
webPageData <- uniprotProteinView::getProtein("Q04206", FALSE)
features <- uniprotProteinView::getFeaturesDataFrame(webPageData)


}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
