% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRetrieval.R
\name{getProtein}
\alias{getProtein}
\title{INTERNAL FUNCTION: Get protein XML data}
\usage{
getProtein(source, showProgress)
}
\arguments{
\item{source}{List of entries for which to attempt to load}

\item{showProgress}{If the user should be notified on the progress of this function}
}
\value{
List for each loaded protein, each element containing a list of XML data
}
\description{
Searches through list or vector of entries and attempts to retrieve UniProt protein
XML data. This can be done by either .xml entry for local files, where if none are found,
will ask if the user either want to skip or download the file and load that. Otherwise, the
user can specify a UniProt protein key_code where the function will then attempt to read
the XML data from the webpage. Alternativly, the user can use "random" to randomly pick
a protein.
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
