% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUniprotProteinView.R
\name{runUniprotProteinDraw}
\alias{runUniprotProteinDraw}
\title{Creates Shiny app}
\usage{
runUniprotProteinDraw()
}
\description{
Will create an interactive shiny app with imbeded plotly figure
}
\references{
TODO references
}
\author{
{George Zorn, \email{george.zorn@mail.utoronto.ca}}
}
